chore: move ruptures components
