+++ ui/app/_components/ruptures/MonthTable.module.css
