+++ ui/app/_components/ruptures/PriorityTable.tsx
