+++ ui/app/_components/ruptures/SearchableTableSection.tsx
